requirements.txt
torch
datasets
accelerate == 0.21.0
peft == 0.4.0
bitsandbytes == 0.40.2
transformers == 4.31.0
trl == 0.4.7
pyyaml
scipy
huggingface-hub
protobuf
sentencepiece
wandb
